Client side statistics (#6)
Added support for collecting client statistics into log. 

The statistics data is collected for an interval of time. At the end of the interval, the stats data is logged in a specified JSON format that can be filtered and parsed. After the logging, the counters are cleared and collection of data resumes.

Collection intervals are aligned to the top of the hour. This means first interval logs may contain stats for a shorter interval.

Collection of stats are controlled by the following environment variables:
      NOSQL_STATS_PROFILE=[none|regular|more|all]
         Specifies the stats profile:
         none - disabled,
         regular - per request: counters, errors, latencies, delays, retries.
            This incurs minimum overhead.
         more - stats above with 95th and 99th percentile latencies.
            This may add 0.5% overhead compared to none stats profile.
         all - stats above with per query information.
            This may add 1% overhead compared to none stats profile.

      NOSQL_STATS_INTERVAL=600 Interval in seconds to log the stats, by default
        is 10 minutes.

      NOSQL_STATS_PRETTY_PRINT=true Option to enable pretty printing of the JSON
        data, default value is false.

Collection of stats can also be used by using the API: `NoSQLHandleConfig.set_stats_profile` or `StatsControl.set_profile`. At runtime stats collection can be enabled selectively by using :py:meth:`StatsControl.start` ond `StatsControl.stop`.